requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
anthropic>=0.58.2
aiofiles>=23.0.0